+++ src/tasks/class-ss-setup-task.php
@@ -174,13 +174,12 @@ class Setup_Task extends Task {
 		$files = new \RecursiveIteratorIterator( new \RecursiveDirectoryIterator( $dir, \RecursiveDirectoryIterator::SKIP_DOTS ), \RecursiveIteratorIterator::CHILD_FIRST );
 
 		foreach ( $files as $fileinfo ) {
-			if ( is_multisite() ) {
-				$lookup = untrailingslashit( $dir ) . DIRECTORY_SEPARATOR . Plugin::SLUG . '-' . get_current_blog_id() . '-';
-				// Don't remove files/folders for a different blog.
-				if ( 0 !== strpos( $fileinfo->getRealPath(), $lookup ) ) {
-					continue;
-				}
+			$can_delete_file = apply_filters( 'ss_can_delete_file', true, $fileinfo );
+
+			if ( ! $can_delete_file ) {
+				continue;
 			}
+
 			if ( $fileinfo->isDir() ) {
 				if ( false === rmdir( $fileinfo->getRealPath() ) ) {
 					return false;
